+++ NEWS.md
@@ -1,3 +1,17 @@
+# cxhull 0.4.0
+
+* New function `hullSummary`, which returns a summary of a 3D triangulated 
+convex hull.
+
+* Now, the `facesColor` argument of `plotConvexHull3d` can be a vector, allowing 
+to assign different colors to the faces. See the README file for examples.
+
+* The `plotConvexHull3d` function has a new argument `palette`; it allows to 
+decorate the faces with a color gradient.
+
+* Updated the README file.
+
+
 # cxhull 0.3.0
 
 * New function `plotConvexHull3d`, to easily plot a 3D convex hull, with a